+++ csatar-plugins/csatar/Plugin.php
@@ -20,6 +20,7 @@ use Session;
 use System\Classes\PluginBase;
 use ValidationException;
 use Validator;
+use Csatar\Csatar\Classes\Validators\CnpValidator;
 
 /**
  * csatar Plugin Information File
@@ -160,6 +161,8 @@ class Plugin extends PluginBase
         });
 
         $this->saveGuestMandateTypeIdsForEveryAssociationToSession();
+
+        Validator::extend('cnp', CnpValidator::class);
     }
 
     /**