+++ csatar-plugins/csatar/models/Association.php
@@ -39,6 +39,7 @@ class Association extends OrganizationBase
      */
     public $fillable = [
         'name',
+        'name_abbreviation',
         'coordinates',
         'contact_name',
         'contact_email',
@@ -49,6 +50,7 @@ class Association extends OrganizationBase
         'ecset_code_suffix',
         'team_fee',
         'currency_id',
+        'personal_identification_number_validator'
     ];
 
     /**
@@ -119,5 +121,9 @@ class Association extends OrganizationBase
     public static function getOrganizationTypeModelNameUserFriendly()
     {
         return Lang::get('csatar.csatar::lang.plugin.admin.association.association');
+    }
+
+    public function getPersonalIdentificationNumberValidatorOptions() {
+        return ['cnp' => 'CNP'];
     }
 }